[Fix] 태그, 진행 방식 검증
- 수정 시 모든 컬럼 옵셔널로 변경

+++ src/modules/project/project.service.ts
@@ -7,11 +7,20 @@ import {
 import { PrismaService } from '../prisma/prisma.service';
 import { GetManyProjectDTO } from './dto/get-project.dto';
 import { PAGE_SIZE } from 'src/constants/pagination';
-import { ProjectDTO } from './dto/project.dto';
+import { CreateProjectDTO } from './dto/create-project.dto';
+import { ModifyProjectDTO } from './dto/modify-project.dto';
+import { MethodService } from '../method/method.service';
+import { PositionTagService } from '../position-tag/position-tag.service';
+import { SkillTagService } from '../skill-tag/skill-tag.service';
 
 @Injectable()
 export class ProjectService {
-  constructor(private prismaService: PrismaService) {}
+  constructor(
+    private prismaService: PrismaService,
+    private methodService: MethodService,
+    private positionTagService: PositionTagService,
+    private skillTagService: SkillTagService,
+  ) {}
 
   async fetchManyProject(dto: GetManyProjectDTO) {
     const skipAmount = (dto.page - 1) * PAGE_SIZE;
@@ -161,7 +170,7 @@ export class ProjectService {
     data,
   }: {
     authorId: number;
-    data: ProjectDTO;
+    data: CreateProjectDTO;
   }) {
     const {
       title,
@@ -218,7 +227,7 @@ export class ProjectService {
   }: {
     authorId: number;
     id: number;
-    data: ProjectDTO;
+    data: ModifyProjectDTO;
   }) {
     const {
       title,
@@ -239,9 +248,14 @@ export class ProjectService {
       throw new ForbiddenException('기획자만 수정 가능합니다.');
     }
 
+    if (methodId) {
+      // 진행 방식 검증
+      await this.methodService.fetchMethod({ id: methodId });
+    }
+
     // 트랜잭션 실행
     return await this.prismaService.$transaction(async (prisma) => {
-      // 1. 프로젝트 정보 업데이트
+      // 프로젝트 정보 업데이트
       const updatedProject = await prisma.project.update({
         where: { id },
         data: {
@@ -257,28 +271,47 @@ export class ProjectService {
         },
       });
 
-      // 2. 기존 태그 삭제
-      await prisma.projectSkillTag.deleteMany({
-        where: { projectId: id },
-      });
-      await prisma.projectPositionTag.deleteMany({
-        where: { projectId: id },
-      });
+      if (skillTagId) {
+        // 스킬 태그 검증
+        const skillTags = await Promise.all(
+          skillTagId.map((id) => this.skillTagService.fetchSkillTag({ id })),
+        );
 
-      // 3. 새로운 태그 생성
-      await prisma.projectSkillTag.createMany({
-        data: skillTagId.map((tagId) => ({
-          projectId: id,
-          skillTagId: tagId,
-        })),
-      });
+        // 기존 스킬 태그 삭제
+        await prisma.projectSkillTag.deleteMany({
+          where: { projectId: id },
+        });
 
-      await prisma.projectPositionTag.createMany({
-        data: positionTagId.map((tagId) => ({
-          projectId: id,
-          positionTagId: tagId,
-        })),
-      });
+        // 새로운 스킬 태그 생성
+        await prisma.projectSkillTag.createMany({
+          data: skillTags.map((tag) => ({
+            projectId: id,
+            skillTagId: tag.id,
+          })),
+        });
+      }
+
+      if (positionTagId) {
+        // 포지션 태그 검증
+        const positionTags = await Promise.all(
+          positionTagId.map((id) =>
+            this.positionTagService.fetchPositionTag({ id }),
+          ),
+        );
+
+        // 기존 포지션 태그 삭제
+        await prisma.projectPositionTag.deleteMany({
+          where: { projectId: id },
+        });
+
+        // 새로운 포지션 태그 생성
+        await prisma.projectPositionTag.createMany({
+          data: positionTags.map((tag) => ({
+            projectId: id,
+            positionTagId: tag.id,
+          })),
+        });
+      }
 
       return updatedProject;
     });
